Reorganize group headers. Change date group header to follow the others
